split out sql adapter and util modules
