+++ corehq/apps/userreports/sql/columns.py
@@ -1,5 +1,6 @@
-from sqlagg import SumWhen
+import hashlib
 import sqlalchemy
+from sqlagg import SumWhen
 from corehq.db import Session
 from corehq.apps.reports.sqlreport import DatabaseColumn
 from fluff import TYPE_STRING
@@ -59,6 +60,27 @@ def get_expanded_column_config(data_source_configuration, column_config, lang):
     return SqlColumnConfig(_expand_column(column_config, vals, lang), warnings=column_warnings)
 
 
+def get_column_name(path):
+    """
+    :param path: xpath from form or case
+    :return: column name for postgres
+
+    Postgres only allows columns up to 63 characters
+    Anyone viewing the table directly will want to know the last parts of the path, not the first parts e.g.
+    this: 'my_long_choice_list_option_1_ABCDEFGH', 'my_long_choice_list_option_2_ABCD1234'
+    not: 'question_group_1_my_long_choice_ABCDEFGH', 'question_group_1_my_long_choice_ABCD1234'
+    """
+    parts = path.split("/")
+
+    def _hash(parts):
+        front = "/".join(parts[:-1])
+        end = parts[-1]
+        return hashlib.sha1('{}_{}'.format(hashlib.sha1(front).hexdigest(), end)).hexdigest()[:8]
+
+    new_parts = path[-54:].split("/")
+    return "_".join(new_parts + [_hash(parts)])
+
+
 def _get_distinct_values(data_source_configuration, column_config, expansion_limit=10):
     """
     Return a tuple. The first item is a list of distinct values in the given
@@ -71,7 +93,7 @@ def _get_distinct_values(data_source_configuration, column_config, expansion_lim
     :param expansion_limit:
     :return:
     """
-    from corehq.apps.userreports.sql import get_indicator_table
+    from corehq.apps.userreports.sql.adapter import get_indicator_table
 
     too_many_values = False
